+++ web/lib/components/flashcards/PageAllSets.jsx
@@ -1,4 +1,4 @@
-import { useState, useContext, useEffect } from 'react'
+import { useState, useContext } from 'react'
 import { useRouter } from 'next/router'
 
 import { Alert } from '../common/Alert'
@@ -7,20 +7,13 @@ import { Button } from '../common/Button'
 import { Dialog } from '../common/Dialog'
 import { Layout } from '../common/Layout'
 import { Breadcrumb } from '../common/Breadcrumb'
-import {
-  ArrowLeftIcon,
-  ArrowRightIcon,
-  PlusIcon,
-  DotsVerticalIcon,
-  TrashIcon,
-  EditIcon,
-} from '../common/Icons'
+import { ArrowLeftIcon, ArrowRightIcon, PlusIcon } from '../common/Icons'
 
 import { useUser } from '../../auth'
 import { useTranslation } from '../../i18n'
 import { ToastContext } from '../../context/Toast'
-import { useDropdown } from '../useDropdown'
 
+import { FlashcardSet } from './FlashcardSet'
 import { useAllSets } from './useAllSets'
 
 const PER_PAGE = 9
@@ -58,143 +51,36 @@ function CreateNewSetDialog({
   )
 }
 
-function EditSetDialog({ set, onOpenChange, onSetUpdate, loading, ...props }) {
-  const [name, setName] = useState(set?.name)
+function PageNavigation({
+  onClickNext,
+  onClickPrev,
+  prevDisabled,
+  nextDisabled,
+}) {
   const { _e } = useTranslation()
 
-  const updateName = (e) => {
-    setName(e.target.value)
-  }
-
-  useEffect(() => {
-    setName(set?.name || '')
-  }, [set])
-
   return (
-    <Dialog
-      title={_e('flashcardset.form.name')}
-      onOpenChange={onOpenChange}
-      {...props}>
-      <div className="p-2">
-        <form className="flex w-80 flex-col gap-2" onSubmit={onSetUpdate(name)}>
-          <div className="grid grid-cols-2 gap-2">
-            <Input name="name" value={name} onChange={updateName} />
-            <Button loading={loading} variant="primary" type="submit">
-              {_e('flashcardset.form.updateName')}
-            </Button>
-          </div>
-        </form>
-      </div>
-    </Dialog>
-  )
-}
-
-function ConfirmSetDeletionModal({ onSetDelete, open, onOpenChange, loading }) {
-  const { _e } = useTranslation()
-
-  const handleClick =
-    (confirm = false) =>
-    async (e) => {
-      e.preventDefault()
-
-      confirm && (await onSetDelete())
-      onOpenChange(false)
-    }
-
-  return (
-    <Dialog
-      title={_e('flashcardset.modal.confirm')}
-      open={open}
-      onOpenChange={onOpenChange}
-      dismissable={false}>
-      <form className="p-2" onSubmit={handleClick(true)}>
-        <p className="mb-5 text-red-500">
-          {_e('flashcardset.modal.doYouWantToDeleteThisSet')}
-        </p>
-        <div className="grid grid-cols-2 gap-2">
-          <Button loading={loading} variant="primary" type="submit">
-            {_e('common.confirm')}
-          </Button>
-          <Button disabled={loading} onClick={handleClick()}>
-            {_e('common.cancel')}
-          </Button>
-        </div>
-      </form>
-    </Dialog>
-  )
-}
-
-function FlashcardSet({ set, openDeleteConfimDialog, openEditDialog }) {
-  const { _e } = useTranslation()
-  const router = useRouter()
-
-  const {
-    isOpen,
-    referenceProps,
-    doToggleDropdown,
-    doCloseDropdown,
-    floatingProps,
-  } = useDropdown({ placement: 'bottom-end' })
-
-  const toggleDropdown = (e) => {
-    e.preventDefault()
-    doToggleDropdown()
-  }
-
-  const navigate = () => router.push(`/flashcards/${set.id}`)
-
-  return (
-    <div className="relative flex aspect-video w-full cursor-pointer flex-col items-center justify-center rounded border shadow hover:shadow-lg hover:shadow-sky-200">
-      <p className="text-bold pointer-events-none text-center text-xl">
-        {set.name}
-      </p>
-      <p className="pointer-events-none mt-1 text-center text-sm text-slate-400">
-        ({set.flashcards.length}&nbsp;{_e('flashcard.flashcards')})
-      </p>
-      <div id="options" className="absolute top-1 right-1 z-20 aspect-square">
-        <div className="relative">
-          <Button
-            onClick={toggleDropdown}
-            onBlur={doCloseDropdown}
-            {...referenceProps()}>
-            <DotsVerticalIcon className="h-6 w-6" />
-          </Button>
-          {isOpen && (
-            <div
-              className="z-30 w-36 rounded border border-slate-200 bg-white p-1 text-gray-800 shadow-lg"
-              {...floatingProps({ right: 0 })}>
-              <ul>
-                <li
-                  className="mb-1 flex flex-row items-center justify-between rounded px-2 py-1 hover:bg-slate-100"
-                  onClick={openEditDialog(set)}>
-                  {_e('flashcardset.dropdown.rename')}{' '}
-                  <EditIcon className="h-4 w-4" />
-                </li>
-                <li
-                  className="mb-1 flex flex-row items-center justify-between rounded px-2 py-1 hover:bg-slate-100"
-                  onClick={openDeleteConfimDialog(set)}>
-                  {_e('flashcardset.dropdown.delete')}{' '}
-                  <TrashIcon className="h-4 w-4 text-red-400" />
-                </li>
-              </ul>
-            </div>
-          )}
-        </div>
-      </div>
-      <div
-        id="overlay"
-        className="absolute top-0 left-0 z-10 h-full w-full bg-transparent"
-        onClick={navigate}></div>
+    <div className="flex items-center justify-center gap-2">
+      <Button
+        onClick={onClickPrev}
+        disabled={prevDisabled}
+        className="inline-flex items-center gap-2">
+        <ArrowLeftIcon />
+        {_e('common.previous')}
+      </Button>
+      <Button
+        onClick={onClickNext}
+        disabled={nextDisabled}
+        className="inline-flex items-center gap-2">
+        {_e('common.next')}
+        <ArrowRightIcon />
+      </Button>
     </div>
   )
 }
 
 export function PageAllSets({ page }) {
   const [isDialogOpen, setIsDialogOpen] = useState(false)
-  const [isEditDialogOpen, setIsEditDialogOpen] = useState(false)
-  const [isConfirmModalOpen, setIsConfirmModalOpen] = useState(false)
-  const [editedSet, setEditedSet] = useState(null)
-
   const [currentPage, setCurrentPage] = useState(Number(page) - 1)
 
   const { _e } = useTranslation()
@@ -232,43 +118,39 @@ export function PageAllSets({ page }) {
     )
   }
 
-  const doCreateNewSet = (name) => async (e) => {
+  const doCreateNewSet = (name) => (e) => {
     e.preventDefault()
 
-    await createNewSet.mutateAsync({ name, userId: user.id })
-
-    if (createNewSet.isSuccess) {
-      notify({ title: _e('flashcardset.createNewSetSuccessfully') })
-      setIsDialogOpen(false)
-    }
+    createNewSet.mutate(
+      { name, userId: user.id },
+      {
+        onSuccess: () => {
+          notify({ title: _e('flashcardset.createNewSetSuccessfully') })
+          setIsDialogOpen(false)
+        },
+      },
+    )
   }
 
-  const doUpdateSet = (name) => async (e) => {
-    e.preventDefault()
+  const doUpdateSet = (set) => (name) =>
+    updateSet.mutate(
+      { id: set.id, name: name, userId: set.user_id },
+      {
+        onSuccess: () => {
+          notify({ title: _e('flashcardset.updateNameSuccessfully') })
+        },
+      },
+    )
 
-    await updateSet.mutateAsync({
-      id: editedSet.id,
-      name: name,
-      userId: editedSet.user_id,
-    })
-
-    if (updateSet.isSuccess) {
-      notify({ title: _e('flashcardset.updateNameSuccessfully') })
-      setIsEditDialogOpen(false)
-      setEditedSet(null)
-    }
-  }
-
-  const doDeleteSet = async () => {
-    await deleteSet.mutateAsync({ id: editedSet.id })
-
-    if (deleteSet.isSuccess) {
-      notify({
-        title: _e('flashcardset.deleteSetSuccessfully'),
-      })
-      setEditedSet(null)
-    }
-  }
+  const doDeleteSet = (id) => () =>
+    deleteSet.mutate(
+      { id },
+      {
+        onSuccess: () => {
+          notify({ title: _e('flashcardset.deleteSetSuccessfully') })
+        },
+      },
+    )
 
   if (isLoading) return <Layout loading />
 
@@ -278,27 +160,15 @@ export function PageAllSets({ page }) {
   ]
 
   const openDialog = () => setIsDialogOpen(true)
-  const openDeleteConfimDialog = (set) => (e) => {
-    setEditedSet(set)
-    setIsConfirmModalOpen(true)
-  }
-  const openEditDialog = (set) => (e) => {
-    setEditedSet(set)
-    setIsEditDialogOpen(true)
-  }
 
   return (
     <Layout>
-      <div className="mb-4 grid w-full grid-cols-2 grid-rows-1">
-        <div className="flex w-full items-center">
-          <Breadcrumb links={links} />
-        </div>
-        <div className="flex w-full flex-row items-center justify-end">
-          <Button className="flex items-center gap-2" onClick={openDialog}>
-            <PlusIcon className="h-5 w-5" />{' '}
-            <span className="">{_e('flashcardset.create')}</span>
-          </Button>
-        </div>
+      <div className="mb-4 flex items-center justify-between">
+        <Breadcrumb links={links} />
+        <Button className="flex items-center gap-2" onClick={openDialog}>
+          <PlusIcon className="h-5 w-5" />{' '}
+          <span className="">{_e('flashcardset.create')}</span>
+        </Button>
       </div>
 
       {/* Create new flashcard set */}
@@ -312,27 +182,12 @@ export function PageAllSets({ page }) {
         <Alert variant="danger">{_e('flashcardset.error.createNewSet')}</Alert>
       )}
 
-      {/* Edit flashcard set */}
-      <EditSetDialog
-        open={isEditDialogOpen}
-        onOpenChange={setIsEditDialogOpen}
-        onSetUpdate={doUpdateSet}
-        set={editedSet}
-        loading={updateSet.isLoading}
-      />
       {updateSet.isError && (
         <Alert variant="danger">
           {_e('flashcardset.error.updateNameFail')}
         </Alert>
       )}
 
-      {/* Delete flashcard set */}
-      <ConfirmSetDeletionModal
-        open={isConfirmModalOpen}
-        onOpenChange={setIsConfirmModalOpen}
-        onSetDelete={doDeleteSet}
-        loading={deleteSet.isLoading}
-      />
       {deleteSet.isError && (
         <Alert variant="danger">{_e('flashcardset.error.deleleSet')}</Alert>
       )}
@@ -343,30 +198,22 @@ export function PageAllSets({ page }) {
             <FlashcardSet
               key={set.id}
               set={set}
-              openDeleteConfimDialog={openDeleteConfimDialog}
-              openEditDialog={openEditDialog}
+              doUpdateSet={doUpdateSet(set)}
+              doDeleteSet={doDeleteSet(set.id)}
+              isUpdateLoading={updateSet.isLoading}
+              isDeleteLoading={deleteSet.isLoading}
             />
           ))}
         </div>
       </div>
 
       {/* Navigators */}
-      <div className="flex items-center justify-center gap-2">
-        <Button
-          onClick={prev}
-          disabled={currentPage === 0}
-          className="inline-flex items-center gap-2">
-          <ArrowLeftIcon />
-          {_e('common.previous')}
-        </Button>
-        <Button
-          onClick={next}
-          disabled={flashcardSets.length < PER_PAGE}
-          className="inline-flex items-center gap-2">
-          {_e('common.next')}
-          <ArrowRightIcon />
-        </Button>
-      </div>
+      <PageNavigation
+        onClickPrev={prev}
+        onClickNext={next}
+        prevDisabled={currentPage === 0}
+        nextDisabled={flashcardSets.length < PER_PAGE}
+      />
     </Layout>
   )
 }
